+++ Sources/HIGThemesSystem/HIGAdminTheme.swift
@@ -28,6 +28,7 @@ public struct HIGAdminTheme: HIGTheme, Sendable {
     public let select: any HIGSelectTokens
     public let autocomplete: any HIGSelectTokens
     public let tagInput: any HIGSelectTokens
+    public let dropZone: any HIGDropZoneTokens
     public let toggle: any HIGToggleTokens
     public let checkbox: any HIGCheckboxTokens
     public let radio: any HIGRadioTokens
@@ -123,6 +124,7 @@ public struct HIGAdminTheme: HIGTheme, Sendable {
         self.select = base.select
         self.autocomplete = base.autocomplete
         self.tagInput = base.tagInput
+        self.dropZone = base.dropZone
         self.toggle = base.toggle
         self.checkbox = base.checkbox
         self.radio = base.radio