+++ Sources/HIGThemesSystem/HIGHighContrastTheme.swift
@@ -23,6 +23,7 @@ public struct HIGHighContrastTheme: HIGTheme, Sendable {
     public let select: any HIGSelectTokens
     public let autocomplete: any HIGSelectTokens
     public let tagInput: any HIGSelectTokens
+    public let dropZone: any HIGDropZoneTokens
     public let toggle: any HIGToggleTokens
     public let checkbox: any HIGCheckboxTokens
     public let radio: any HIGRadioTokens
@@ -102,6 +103,7 @@ public struct HIGHighContrastTheme: HIGTheme, Sendable {
         self.select = base.select
         self.autocomplete = base.autocomplete
         self.tagInput = base.tagInput
+        self.dropZone = base.dropZone
         self.toggle = base.toggle
         self.checkbox = base.checkbox
         self.radio = base.radio